api_testing/requirements.txt
attrs==20.1.0
configparser==4.0.2
contextlib2==0.6.0.post1
importlib-metadata==1.7.0
jsonschema==3.2.0
pathlib2==2.3.5
pyrsistent==0.16.0
robotframework==3.2.1
scandir==1.10.0
six==1.15.0
zipp==1.2.0
pyyaml
robotframework-requests